Some refactory for Cipher class
