+++ src/main/java/com/intel/chimera/crypto/Cipher.java
@@ -43,87 +43,6 @@ public abstract class Cipher {
   public static final int DECRYPT_MODE = 0;
 
   /**
-   * Get a cipher instance for specified algorithm/mode/padding.
-   *
-   * @param props
-   *          the configuration properties
-   * @param transformation
-   *          algorithm/mode/padding
-   * @return Cipher the cipher. Null value will be returned if no
-   *         cipher classes with transformation configured.
-   */
-  public static Cipher getInstance(Properties props,
-      CipherTransformation transformation) throws GeneralSecurityException  {
-    List<Class<? extends Cipher>> klasses =
-        getCipherClasses(props, transformation);
-    Cipher cipher = null;
-    if (klasses != null) {
-      for (Class<? extends Cipher> klass : klasses) {
-        try {
-          cipher = ReflectionUtils.newInstance(klass, props, transformation);
-          if(cipher != null) {
-            LOG.debug("Using cipher {} for transformation {}.", klass.getName(), transformation.getName());
-            break;
-          }
-        } catch (Exception e) {
-          LOG.debug("Cipher {} is not available or transformation {} is not supported.",
-              klass.getName(), transformation.getName());
-        }
-      }
-    }
-
-    return (cipher == null)?new JceCipher(props, transformation):cipher;
-  }
-
-  /**
-   * Get a cipher for algorithm/mode/padding in config value
-   * chimera.crypto.cipher.transformation
-   *
-   * @return Cipher the cipher object Null value will be returned if no
-   *         cipher classes with transformation configured.
-   */
-  public static Cipher getInstance() throws GeneralSecurityException {
-    return getInstance(new Properties());
-  }
-
-  /**
-   * Get a cipher for algorithm/mode/padding in config value
-   * chimera.crypto.cipher.transformation
-   *
-   * @param props the properties which contain the configurations
-   *         of the crypto cipher
-   * @return Cipher the cipher object Null value will be returned if no
-   *         cipher classes with transformation configured.
-   */
-  public static Cipher getInstance(Properties props)
-      throws GeneralSecurityException {
-    return getInstance(props, Utils.getCripherTransformation(props));
-  }
-
-  private static List<Class<? extends Cipher>> getCipherClasses(
-      Properties props, CipherTransformation transformation) {
-    List<Class<? extends Cipher>> result = Lists.newArrayList();
-    String cipherClassString = Utils.getCipherClassString(props);
-    if (cipherClassString == null) {
-      LOG.debug("No cipher classes configured.");
-      return null;
-    }
-    for (String c : Splitter.on(',').trimResults().omitEmptyStrings().
-        split(cipherClassString)) {
-      try {
-        Class<?> cls = ReflectionUtils.getClassByName(c);
-        result.add(cls.asSubclass(Cipher.class));
-      } catch (ClassCastException e) {
-        LOG.error("Class {} is not a Cipher.", c);
-      } catch (ClassNotFoundException e) {
-        LOG.error("Cipher {} not found.", c);
-      }
-    }
-
-    return result;
-  }
-
-  /**
    * @return the CipherTransformation for this cipher.
    */
   public abstract CipherTransformation getTransformation();
